backend/requirements.txt
Django
langchain
langchain-classic
langchain-community
langchain-core
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
Markdown
numpy
ollama
pillow
python-dotenv
requests
gunicorn
psycopg2-binary
redis
django-redis
langchain-ollama
dj-database-url
pypdf
boto3
django-storages
